+++ timesketch/api/v1/routes.py
@@ -15,46 +15,42 @@
 
 from __future__ import unicode_literals
 
-from .resources import AggregationGroupResource
-from .resources import AggregationGroupListResource
-from .resources import AggregationListResource
-from .resources import AggregationLegacyResource
-from .resources import AggregationExploreResource
-from .resources import AggregationInfoResource
-from .resources import AggregationResource
-from .resources import AnalysisResource
-from .resources import AnalyzerRunResource
-from .resources import AnalyzerSessionResource
-from .resources import ExploreResource
-from .resources import EventResource
-from .resources import EventAnnotationResource
-from .resources import EventCreateResource
-from .resources import EventTaggingResource
-from .resources import GraphResource
-from .resources import GraphViewListResource
-from .resources import GraphViewResource
-from .resources import SketchResource
-from .archive import SketchArchiveResource
-from .resources import SketchListResource
-from .resources import ViewResource
-from .resources import ViewListResource
-from .resources import SearchTemplateResource
-from .resources import SearchTemplateListResource
-from .resources import UploadFileResource
-from .resources import TaskResource
-from .resources import StoryListResource
-from .resources import StoryResource
-from .resources import QueryResource
-from .resources import CountEventsResource
-from .resources import TimelineResource
-from .resources import TimelineListResource
-from .resources import SearchIndexListResource
-from .resources import SearchIndexResource
-from .resources import SessionResource
-from .resources import UserListResource
-from .resources import GroupListResource
-from .resources import CollaboratorResource
-from .resources import LoggedInUserResource
+from .resources.aggregation import AggregationGroupResource
+from .resources.aggregation import AggregationGroupListResource
+from .resources.aggregation import AggregationListResource
+from .resources.aggregation import AggregationExploreResource
+from .resources.aggregation import AggregationInfoResource
+from .resources.aggregation import AggregationResource
+from .resources.analysis import AnalysisResource
+from .resources.analysis import AnalyzerRunResource
+from .resources.analysis import AnalyzerSessionResource
+from .resources.explore import ExploreResource
+from .resources.event import EventResource
+from .resources.event import EventAnnotationResource
+from .resources.event import EventCreateResource
+from .resources.event import EventTaggingResource
+from .resources.event import CountEventsResource
+from .resources.sketch import SketchResource
+from .resources.sketch import SketchListResource
+from .resources.archive import SketchArchiveResource
+from .resources.view import ViewResource
+from .resources.view import ViewListResource
+from .resources.searchtemplate import SearchTemplateResource
+from .resources.searchtemplate import SearchTemplateListResource
+from .resources.upload import UploadFileResource
+from .resources.task import TaskResource
+from .resources.story import StoryListResource
+from .resources.story import StoryResource
+from .resources.explore import QueryResource
+from .resources.timeline import TimelineResource
+from .resources.timeline import TimelineListResource
+from .resources.searchindex import SearchIndexListResource
+from .resources.searchindex import SearchIndexResource
+from .resources.session import SessionResource
+from .resources.user import UserListResource
+from .resources.user import GroupListResource
+from .resources.user import CollaboratorResource
+from .resources.user import LoggedInUserResource
 
 
 # Disable error for long line. Readability is more important than line
@@ -70,7 +66,6 @@ API_ROUTES = [
     (AggregationListResource, '/sketches/<int:sketch_id>/aggregation/'),
     (AggregationGroupResource, '/sketches/<int:sketch_id>/aggregation/group/<int:group_id>/'),
     (AggregationGroupListResource, '/sketches/<int:sketch_id>/aggregation/group/'),
-    (AggregationLegacyResource, '/sketches/<int:sketch_id>/aggregation/legacy/'),
     (AggregationExploreResource, '/sketches/<int:sketch_id>/aggregation/explore/'),
     (AggregationInfoResource, '/aggregation/info/'),
     (AggregationResource, '/sketches/<int:sketch_id>/aggregation/<int:aggregation_id>/'),
@@ -93,9 +88,6 @@ API_ROUTES = [
     (TimelineResource, '/sketches/<int:sketch_id>/timelines/<int:timeline_id>/'),
     (SearchIndexListResource, '/searchindices/'),
     (SearchIndexResource, '/searchindices/<int:searchindex_id>/'),
-    (GraphResource, '/sketches/<int:sketch_id>/explore/graph/'),
-    (GraphViewListResource, '/sketches/<int:sketch_id>/explore/graph/views/'),
-    (GraphViewResource, '/sketches/<int:sketch_id>/explore/graph/views/<int:view_id>/'),
     (SessionResource, '/sketches/<int:sketch_id>/explore/sessions/<string:timeline_index>'),
     (UserListResource, '/users/'),
     (GroupListResource, '/groups/'),
